Use type instead of metadata file for cpe_processing_rules
